Move classes related to fetching records to own package CIRC-624
