+++ src/main/java/org/folio/circulation/support/FindWithCqlQuery.java
@@ -4,8 +4,12 @@ import java.util.concurrent.CompletableFuture;
 
 import org.folio.circulation.domain.MultipleRecords;
 import org.folio.circulation.support.http.client.CqlQuery;
+import org.folio.circulation.support.http.client.PageLimit;
 
 public interface FindWithCqlQuery<T> {
     CompletableFuture<Result<MultipleRecords<T>>> findByQuery(
       Result<CqlQuery> queryResult);
+
+  CompletableFuture<Result<MultipleRecords<T>>> findByQuery(
+    Result<CqlQuery> queryResult, PageLimit pageLimit);
 }
